Create a scanner and tests
We finally add the first pass of our compiler for the VM by adding
a scanner. This reads in a source file and generates the tokens we need
to compile things later. We also setup the beginning of the test suite
by adding tests to our codebase. In order to test individual passes we
split the compiler into a main.rs and lib.rs file so that we have access
to all of the internal code, while main.rs acts as a wrapper the lib.rs
main method. This means that later we could setup test code to validate
the output of calling the compiler and test it's individual parts in the
same test suite, rather than having unit tests everywhere in the source.