Implement parser and compiler for Cedar exprs
This parser uses Vaughan Pratt's 'top-down operator precedence parsing'
method in order to parse the tokens and then emit the bytecode into the
Chunk. This means we now have a full bytecode interpreter from beginning
to end. Overtime we'll begin adding in all the other possible statements
and expressions to the parser and compiler!