Add comparison and negate ops, and bool/null types
We continue our work on extending the compiler by adding instructions
for boolean operations, negations, and extending the types to be bool
and null. While the book has a notion of truthiness, we want to
eventually head towards static types and so only allow bools to do
comparcomparisons with each other for now. Later the idea of Ordering
can be extended to things like strings with other strings for instance.