Implement Strings and a small garbage collector
We finally add Strings to the language and what operations you can do on
them into the language. As part of the implementation we create a heap
for the VM and the concept of a heap pointer for the stack machine. For
now the heap will just grow until it gets garbage collected. This is
fine as later we will implement an actual garbage collector, we just
need to deal with the fact that we need a place to store values when not
in use or need longer term storage like variables.