Clean up native function code and make a libstd
If we want or language to be useful then it needs to have I/O! This
commit starts this, by creating a library to simply read and write
contents to a file. We'll need to add more functions from the Rust
standard library for the various types as well, but we're on our way to
having a more full featured language rather than one that is "pure" so
to speak and has no side effects beyond printing values out. We also
fixed up some of the code for loading in libstd to the vm itself. From
here we can finish up the rest of the book and work on expanding Cedar's
capabilities as well as giving it some actual tooling beyond just
a compiler. First class dev experience is really important!