Add better testing for cedar
Up to this point we really weren't testing cedar running any programs.
This changes that by adding an `e2e` test file that invokes the binary
on the actual scripts and checks that the output is what we would
expect. This is nice for things that it won't matter what the internal
implementation is, just that the outside works as expected. We can also
expand testing for internal implementations and thus can have these
tests be independent of each other, but also complimentary. We check
that we can't break user code and we check that our internal assumptions
don't break either.